requirements.txt
sqlalchemy
pandas
python-dotenv
psycopg2-binary